+++ src/Application/Application.RuntimeProfileManager.psm1
@@ -1,6 +1,6 @@
 Set-StrictMode -Version Latest
 
-$script:HybridRuntimeProfileManagerVersion = 'v0.8.2'
+$script:HybridRuntimeProfileManagerVersion = 'v0.8.3'
 
 function Add-HybridRuntimeProfileManagerTypeMetadata {
     param(